streamlit/requirements.txt
IMDbPY==2022.7.9
IMDbPY==2022.7.9
pandas==1.0.5
pandas==2.0.3
Pillow==7.2.0
Pillow==10.3.0
Requests==2.31.0
torch==2.2.2
transformers==4.35.0
transformers==4.40.2
